requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.1
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.1.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.9.0
cachetools==6.2.2
category_encoders==2.8.1
celery==5.3.4
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
comtypes==1.4.13
contourpy==1.3.2
cramjam==2.11.0
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.78.0
dateparser==1.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
email-validator==2.0.0.post2
executing==2.2.0
fastapi==0.128.0
fastparquet==2025.12.0
filelock==3.19.1
Flask==2.3.3
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.1.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.58.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.2.1
google-genai==1.56.0
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huepy==1.2.1
huey==2.6.0
idna==3.10
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
joblib==1.5.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kombu==5.5.4
langdetect==1.0.9
lazy_loader==0.4
librosa==0.11.0
lightgbm==4.6.0
livereload==2.7.1
llvmlite==0.45.1
Mako==1.3.10
Markdown==3.9
markdown2==2.5.4
MarkupSafe==2.1.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
mne==1.10.1
mne-bids==0.17.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.3
narwhals==2.16.0
nest-asyncio==1.6.0
networkx==3.5
nodejs==0.1.1
numba==0.62.1
numpy==2.2.6
opencv-python==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optional-django==0.1.0
optuna==4.5.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.8
playsound==1.2.2
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.3
pypdf==6.6.2
pypiwin32==223
pytest==8.4.2
python-binance==1.0.29
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.99
pytz==2025.2
pywin32==310
PyYAML==6.0.3
pyzmq==26.4.0
redis==5.0.1
referencing==0.37.0
regex==2024.11.6
requests==2.31.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SpeechRecognition==3.14.4
SQLAlchemy==1.4.46
sqlparse==0.5.5
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.50.0
statsmodels==0.14.4
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.8.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
twilio==8.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.40.0
vine==5.1.0
vite==1.5.2
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.3.7
WTForms==3.0.1
xgboost==3.0.2
xlsxwriter==3.2.9
yarl==1.20.1
zipp==3.23.0
